backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy==1.26.2
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.0